requirements.txt
yt-dlp>=2023.12.30
colorama>=0.4.6
ffmpeg-python>=0.2.0
